Addition of conflict and disaster widgets
- Addition of third widget
- Refactor of country profile and error handling
